Initial support for instruction functions
Begin emitting instruction function information:
```
{
  "mnemonic": "c.bnez",
  ...
  "function": "execute(BTYPE((EXTS(13, bitvector_concat(imm, 0b0)), zreg, creg2reg_idx(rs), RISCV_BNE)))",
}
```

Note that the content is in some sort of post-processed form.
The preference would be to match as close as possible what is in the
specification.
```
function clause execute (C_BNEZ(imm, rs)) =
  execute(BTYPE(EXTS(imm @ 0b0), zreg, creg2reg_idx(rs), RISCV_BNE))
```

More careful parsing may be required.

Also, throw in JSON fields "name" and "description" and some wishful thinking.

Signed-off-by: Paul A. Clarke <pclarke@ventanamicro.com>